Refactor namespaces and reorder using directives
Changed the namespace for `IntrusionDetectionRecord.cs` and `IntrusionDetectionReportHandler.cs` to `Genetec.Dap.CodeSamples.Server.ReportHandlers.Intrusion` for better modularization of intrusion-related components. Reordered `using` directives in `IntrusionDetectionReportHandler.cs` for consistency and adherence to coding style guidelines.
