+++ Samples/Plugin SDK/CustomReportSample/Server/ReportHandlers/Intrusion/IntrusionDetectionRecord.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright 2025 Genetec Inc.
 // Licensed under the Apache License, Version 2.0
 
-namespace Genetec.Dap.CodeSamples.Server.ReportHandlers;
+namespace Genetec.Dap.CodeSamples.Server.ReportHandlers.Intrusion;
 
 using System;
